requirements.txt
Flask==3.1.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.41
PyMySQL==1.1.1
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.2.1
python-dotenv~=1.1.0
pip~=25.1.1
attrs~=24.2.0
pillow~=10.4.0
typing_extensions~=4.12.2
MarkupSafe~=3.0.2
blinker~=1.9.0
greenlet~=3.2.2
psycopg2~=2.9.9
cryptography~=45.0.3